Split off gceurlmap into its own file
