+++ pkg/utils/gceurlmap.go
@@ -1,0 +1,65 @@
+/*
+Copyright 2017 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package utils
+
+import (
+	"fmt"
+
+	compute "google.golang.org/api/compute/v1"
+)
+
+// GCEURLMap is a nested map of hostname->path regex->backend
+type GCEURLMap map[string]map[string]*compute.BackendService
+
+// GetDefaultBackend performs a destructive read and returns the default
+// backend of the urlmap.
+func (g GCEURLMap) GetDefaultBackend() *compute.BackendService {
+	var d *compute.BackendService
+	var exists bool
+	if h, ok := g[DefaultBackendKey]; ok {
+		if d, exists = h[DefaultBackendKey]; exists {
+			delete(h, DefaultBackendKey)
+		}
+		delete(g, DefaultBackendKey)
+	}
+	return d
+}
+
+// String implements the string interface for the GCEURLMap.
+func (g GCEURLMap) String() string {
+	msg := ""
+	for host, um := range g {
+		msg += fmt.Sprintf("%v\n", host)
+		for url, be := range um {
+			msg += fmt.Sprintf("\t%v: ", url)
+			if be == nil {
+				msg += fmt.Sprintf("No backend\n")
+			} else {
+				msg += fmt.Sprintf("%v\n", be.Name)
+			}
+		}
+	}
+	return msg
+}
+
+// PutDefaultBackend performs a destructive write replacing the
+// default backend of the url map with the given backend.
+func (g GCEURLMap) PutDefaultBackend(d *compute.BackendService) {
+	g[DefaultBackendKey] = map[string]*compute.BackendService{
+		DefaultBackendKey: d,
+	}
+}
