+++ pkg/utils/gceurlmap_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2015 The Kubernetes Authors.
+Copyright 2017 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,5 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package utils contains shared data structures and functions.
 package utils
+
+import (
+	"testing"
+)
+
+func TestGCEURLMap(t *testing.T) {
+	// TODO
+}
